[Orc] Add program name overriding to OrcRPCTPCServer, llvm-jitlink-executor.
This allows OrcRPCTPCServer clients to set a program name to be used as argv[0]
when running target functions via runAsMain.

If not set the argv vector is passed to the target function unmodified. In this
case JIT-side callers (TargetProcessControl client) should supply the program
name as the first argument if one is expected.

llvm-jitlink-executor is updated to set the program name to
"llvm-jitlink-executor".