[ORC] Runtime work-in-progress. Take 2.
Adds a compiler-rt library (compiler-rt/lib/orc) that includes support for
cxa_atexit interposing; dlfcn.h function emulation; and a new runtime function,
__orc_rt_macho_run_program, for executing JIT'd programs by dlopening the
Main JITDylib, dlsym'ing main, then dlclosing the JITDylib. Emulation of the
dlfcn functions includes running static constructors and destructors.

The llvm-jitlink tool is updated to automatically load the runtime on Darwin and
use it to run target programs. The effect of this is that objects added to
llvm-jitlink (at least on Darwin) will now have their static constructors and
destructors executed, both in in-process mode and out-of-process mode.