[JITLink] Introduce virtual offsets for Symbol definitions in LinkGraph
On AArch32 branch offsets explicitly encode the target instruction subset (Arm/Thumb) in their least significant bit. We want to store Thumb symbol definitions with the extra bit set, so that branch offset calculations account for instruction subset changes automatically.

However, we don't want to adjust the physical size of the symbol or the containing block, since this bit is not part of the symbol's physical location. Instead, we could consider it a virtual offset: Virtual offsets do affect branch offsets, but not physical memory size.

Differential Revision: https://reviews.llvm.org/D144714