[lldb][gnustep] Add basic test and infrastructure for GNUstep ObjC runtime
We are planning to add basic ObjC debug support for the GNUstep ObjC runtime on Linux and Windows: https://github.com/gnustep/libobjc2 On Linux much of the desired functionality is there already. It seems accidental in parts, but it's a good early baseline. On Windows nothing works yet.

In this first patch I want to add the necessary test infrastructure to utilize the GNUstep runtime in the LLDB test suite. The minimal test demonstrates that it works on Linux. The next goal is to extend the PDB parser and get Windows to this baseline as well.

Differential Revision: https://reviews.llvm.org/D146058